Moving classes to the right places
