+++ src/main/java/br/unb/cic/mop/eh/ErrorCollector.java
@@ -1,4 +1,4 @@
-package br.unb.cic.mop.jca.eh;
+package br.unb.cic.mop.eh;
 
 
 import java.util.ArrayList;
