+++ src/main/mop/KeyPairGeneratorSpec.mop
@@ -3,8 +3,8 @@ package mop;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 
-import br.unb.cic.mop.jca.eh.*;
-import br.unb.cic.mop.jca.util.ExecutionContext;
+import br.unb.cic.mop.eh.*;
+import br.unb.cic.mop.ExecutionContext;
 
 /**
  * KeyPairGeneratorSpec
