+++ src/main/mop/KeyStoreSpec.mop
@@ -6,7 +6,8 @@ import java.security.KeyStore.ProtectionParameter;
 import java.security.KeyStore.Entry;
 import java.security.Key key;
 
-import br.unb.cic.mop.jca.util.ExecutionContext;
+import br.unb.cic.mop.eh.*;
+import br.unb.cic.mop.ExecutionContext;
 
 /**
  * KeyStoreSpec
