+++ src/test/java/CipherInputStreamTest.java
@@ -1,6 +1,5 @@
-import br.unb.cic.mop.jca.eh.ErrorCollector;
+import br.unb.cic.mop.eh.ErrorCollector;
 import br.unb.cic.mop.test.Assertions;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
