+++ src/test/java/CipherTest.java
@@ -1,6 +1,6 @@
-import br.unb.cic.mop.jca.eh.ErrorCollector;
+import br.unb.cic.mop.eh.ErrorCollector;
 
-import br.unb.cic.mop.jca.util.ExecutionContext;
+import br.unb.cic.mop.ExecutionContext;
 import br.unb.cic.mop.test.Assertions;
 import org.junit.Before;
 import org.junit.Ignore;
